+++ evalml/tests/pipeline_tests/classification_pipeline_tests/test_classification.py
@@ -2,7 +2,6 @@ from itertools import product
 
 import pandas as pd
 import pytest
-from pandas.testing import assert_series_equal
 
 
 @pytest.mark.parametrize("problem_type", ["binary", "multi"])
@@ -62,13 +61,11 @@ def test_pipeline_has_classes_property(
         else:
             answer = ["class_0", "class_1", "class_2"]
 
-    with pytest.raises(
-        AttributeError, match="Cannot access class names before fitting the pipeline."
-    ):
-        pipeline.classes_
+    # Check that .classes_ is None before fitting
+    assert pipeline.classes_ is None
 
     pipeline.fit(X, y)
-    assert_series_equal(pd.Series(pipeline.classes_), pd.Series(answer))
+    assert pipeline.classes_ == answer
 
 
 def test_woodwork_classification_pipeline(