+++ evalml/pipelines/pipeline_base.py
@@ -254,7 +254,7 @@ class PipelineBase(ABC, metaclass=PipelineBaseMeta):
             pipeline_dict.update({"components": component_dict})
             return pipeline_dict
 
-    def compute_estimator_features(self, X, y=None, X_train=None, y_train=None):
+    def transform_all_but_final(self, X, y=None, X_train=None, y_train=None):
         """Transforms the data by applying all pre-processing components.
 
         Args:
@@ -266,7 +266,7 @@ class PipelineBase(ABC, metaclass=PipelineBaseMeta):
         Returns:
             pd.DataFrame: New transformed features.
         """
-        return self.component_graph.compute_final_component_features(X, y=y)
+        return self.component_graph.transform_all_but_final(X, y=y)
 
     def _fit(self, X, y):
         self.input_target_name = y.name