+++ docs/source/release_notes.rst
@@ -8,6 +8,7 @@ Release Notes
         * Pinned ``cmdstan`` to ``0.28.0`` in ``cmdstan-builder`` to prevent future breaking of support for Prophet :pr:`2880`
         * Added ``Jarque-Bera`` to the ``TargetDistributionDataCheck`` :pr:`2891`
     * Changes
+        * Updated pipelines to use a label encoder component instead of doing encoding on the pipeline level :pr:`2821`
         * Deleted scikit-learn ensembler :pr:`2819`
         * Refactored pipeline building logic out of ``AutoMLSearch`` and into ``IterativeAlgorithm`` :pr:`2854`
         * Refactored names for methods in ``ComponentGraph`` :pr:`2902`
@@ -21,9 +22,11 @@ Release Notes
 .. warning::
 
     **Breaking Changes**
+        * Updated pipelines to use a label encoder component instead of doing encoding on the pipeline level. This means that pipelines will no longer automatically encode non-numerical targets. Please use a label encoder if working with classification problems and non-numeric targets. :pr:`2821`
         * Deleted scikit-learn ensembler :pr:`2819`
         * ``IterativeAlgorithm`` now requires X, y, problem_type as required arguments as well as sampler_name, allowed_model_families, allowed_component_graphs, max_batches, and verbose as optional arguments :pr:`2854`
         * Changed method names of ``fit_features`` and ``compute_final_component_features`` to ``fit_and_transform_all_but_final`` and ``transform_all_but_final`` in ``ComponentGraph`` :pr:`2902`
+
 
 **v0.34.0 Sep. 30, 2021**
     * Enhancements